+++ je_auto_control/utils/change_localize/__init__.py
@@ -1,0 +1,6 @@
+"""Attribute a screen change to the specific element boxes that changed."""
+from je_auto_control.utils.change_localize.change_localize import (
+    localize_changes, rank_changes,
+)
+
+__all__ = ["localize_changes", "rank_changes"]